Exercises on DevNet.
Ok, now the exercises generator is working, later i'll upload a better and more compleate readme.
For now the exercise generator, create the `json` files, upload the `JSONs` to `nft.storage` and then with the `CIDs` of the exercises, we create the exercises in the `DevNet` of Solana.
It would only be necessary to implement listen Exercise